feat(cart-app) Hooks personalizado
implement Hooks personalizado
